@@ -1,28 +1,23 @@
 #!/usr/bin/env python3
-
 import argparse
-import glob
 import os
-import shutil
 import sys
-from pathlib import Path
 
-sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
-import common as C
+import ctp_build
+import ctp_package
+import ctp_test
+import java_common
 
-import build as JB
-import package as JP
-import test as JT
-import javaCommon as JC
-import installLibs as L
+sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
+import common  # noqa: E402
 
 
 def main() -> int:
 
     parser = argparse.ArgumentParser()
-    C.add_clean_arg(parser)
-    C.add_tag_arg(parser)
-    JC.add_common_args(parser)
+    common.add_clean_arg(parser)
+    common.add_tag_arg(parser)
+    java_common.add_common_args(parser)
     parser.add_argument(
         "--skip-tests",
         action="store_true",
@@ -41,17 +36,17 @@ def main() -> int:
 
     # Build and test
     if args.skip_tests:
-        rc = JB.main(build_args)
+        rc = ctp_build.main(build_args)
     else:
         if args.skip_integration_tests:
             build_args.append("-PunitTests")
-        rc = JT.main(build_args)
+        rc = ctp_test.main(build_args)
 
     if rc:
         return rc
 
     # Package
-    rc = JP.main((args.tag,))
+    rc = ctp_package.main((args.tag,))
 
     return 0
 
