@@ -1,17 +1,14 @@
 #!/usr/bin/env python3
-
 import argparse
 import os
-import platform
 import shutil
 import sys
-from typing import Optional
-from typing import Sequence
+from collections.abc import Sequence
 
-sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
-import common as C
+import dotnet_common
 
-import dotnetCommon as DC
+sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
+import common  # noqa: E402
 
 _LINUX = "linux"
 _WINDOWS = "win"
@@ -25,50 +22,54 @@ def runtime_platform() -> str:
     raise ValueError(os.name)
 
 
-def main(argv: Optional[Sequence[str]] = None) -> int:
+def main(argv: Sequence[str] | None = None) -> int:
 
     parser = argparse.ArgumentParser()
-    C.add_clean_arg(parser)
-    C.add_tag_arg(parser)
-    DC.add_args(parser)
+    common.add_clean_arg(parser)
+    common.add_tag_arg(parser)
+    dotnet_common.add_args(parser)
     parser.add_argument(
         "--no-build",
         action="store_true",
     )
     args = parser.parse_args(argv)
 
-    dist_tag_dir = C.DIST_DIR / args.tag
+    dist_tag_dir = f"{common.DIST_DIR}/{args.tag}"
 
-    if args.clean and dist_tag_dir.is_dir():
+    if args.clean and os.path.isdir(dist_tag_dir):
         shutil.rmtree(dist_tag_dir)
-    dist_tag_dir.mkdir(parents=True, exist_ok=True)
+    os.makedirs(dist_tag_dir, exist_ok=True)
 
     platform = runtime_platform()
     rid = f"{platform}-x64"
     smi_services_output_dir = f"smi-services-{args.tag}-{rid}"
 
-    cmd = (
+    cmd: tuple[str, ...] = (
         "dotnet",
         "publish",
         "-warnaserror",
         "--use-current-runtime",
-        "--configuration", args.configuration,
+        "--configuration",
+        args.configuration,
         "--no-build" if args.no_build else "",
         "-p:PublishTrimmed=false",
         "--self-contained",
-        "--output", dist_tag_dir / smi_services_output_dir,
-        "--verbosity", "quiet",
+        "--output",
+        f"{dist_tag_dir}/{smi_services_output_dir}",
+        "--verbosity",
+        "quiet",
         "--nologo",
         "src/SmiServices",
     )
-    C.run(cmd)
+    common.run(cmd)
 
     if platform == _LINUX:
         cmd = (
             "tar",
-            "-C", dist_tag_dir,
+            "-C",
+            dist_tag_dir,
             "-czf",
-            dist_tag_dir / f"{smi_services_output_dir}.tgz",
+            f"{dist_tag_dir}/{smi_services_output_dir}.tgz",
             smi_services_output_dir,
         )
     elif platform == _WINDOWS:
@@ -80,17 +81,17 @@ def main(argv: Optional[Sequence[str]] = None) -> int:
             "-tzip",
             "-mx9",
             "-r",
-            dist_tag_dir / f"{smi_services_output_dir}.zip",
+            f"{dist_tag_dir}/{smi_services_output_dir}.zip",
             smi_services_output_dir,
         )
     else:
         print(f"Error: No case for platform {platform}", file=sys.stderr)
         return 1
 
-    C.run(cmd)
-    shutil.rmtree(dist_tag_dir / smi_services_output_dir)
+    common.run(cmd)
+    shutil.rmtree(f"{dist_tag_dir}/{smi_services_output_dir}")
 
-    C.create_checksums(dist_tag_dir, "smiservices")
+    common.create_checksums(dist_tag_dir, "smiservices")
 
     return 0
 
