@@ -13,10 +13,10 @@ namespace SmiServices.Common.Options
             return options;
         }
 
-        private MongoDbOptions SetMongoPassword(MongoDbOptions opt)
+        private static MongoDbOptions SetMongoPassword(MongoDbOptions opt)
         {
             //get the environment variables current value
-            string? envVar = Environment.GetEnvironmentVariable("MONGO_SERVICE_PASSWORD");
+            var envVar = Environment.GetEnvironmentVariable("MONGO_SERVICE_PASSWORD");
 
             //if there's an env var for it and there are mongodb options being used
             if (!string.IsNullOrWhiteSpace(envVar))
