@@ -1,6 +1,5 @@
 
 using NUnit.Framework;
-using NUnit.Framework.Internal;
 using RabbitMQ.Client;
 using RabbitMQ.Client.Exceptions;
 using System;
@@ -12,54 +11,42 @@ namespace SmiServices.IntegrationTests
 {
     [AttributeUsage(AttributeTargets.Method | AttributeTargets.Class | AttributeTargets.Interface |
                     AttributeTargets.Assembly, AllowMultiple = true)]
-    public class RequiresRabbit : RequiresExternalService
+    public sealed class RequiresRabbit : RequiresExternalService
     {
-        protected override void ApplyToContextImpl(TestExecutionContext context)
-        {
+        public static readonly Lazy<IConnection> Connection = new(GetConnectionFactory);
 
-            var factory = GetConnectionFactory();
+        protected override string? ApplyToContextImpl()
+        {
+            try
+            {
+                CheckExchange();
+                return null;
+            }
+            catch (BrokerUnreachableException e)
+            {
+                return $"Could not connect to RabbitMQ{Environment.NewLine}{e.Message}";
+            }
+        }
+
+        private static IConnection GetConnectionFactory()
+        {
+            var deserializer = new DeserializerBuilder()
+                .IgnoreUnmatchedProperties()
+                .Build();
+            var factory = deserializer.Deserialize<ConnectionFactory>(
+                new StreamReader(Path.Combine(TestContext.CurrentContext.TestDirectory, "Rabbit.yaml")));
             factory.ContinuationTimeout = TimeSpan.FromSeconds(5);
             factory.HandshakeContinuationTimeout = TimeSpan.FromSeconds(5);
             factory.RequestedConnectionTimeout = TimeSpan.FromSeconds(5);
             factory.SocketReadTimeout = TimeSpan.FromSeconds(5);
             factory.SocketWriteTimeout = TimeSpan.FromSeconds(5);
-
-            try
-            {
-                using var conn = factory.CreateConnection();
-                using var model = conn.CreateModel();
-                model.ExchangeDeclare("TEST.ControlExchange", ExchangeType.Topic, durable: true);
-            }
-            catch (BrokerUnreachableException e)
-            {
-                StringBuilder sb = new();
-
-                sb.AppendLine($"Uri:         {factory.Uri}");
-                sb.AppendLine($"Host:        {factory.HostName}");
-                sb.AppendLine($"VirtualHost: {factory.VirtualHost}");
-                sb.AppendLine($"UserName:    {factory.UserName}");
-                sb.AppendLine($"Port:        {factory.Port}");
-
-                string msg = $"Could not connect to RabbitMQ {Environment.NewLine}{sb}{Environment.NewLine}{e.Message}";
-
-                // NOTE(rkm 2021-01-30) Don't fail for Windows CI builds
-                bool shouldFail = FailIfUnavailable && !Environment.OSVersion.ToString().Contains("windows", StringComparison.CurrentCultureIgnoreCase);
-
-                if (shouldFail)
-                    Assert.Fail(msg);
-                else
-                    Assert.Ignore(msg);
-            }
+            return factory.CreateConnection();
         }
 
-        public static ConnectionFactory GetConnectionFactory()
+        public void CheckExchange()
         {
-            IDeserializer deserializer = new DeserializerBuilder()
-                .IgnoreUnmatchedProperties()
-                .Build();
-
-            return deserializer.Deserialize<ConnectionFactory>(new StreamReader(Path.Combine(TestContext.CurrentContext.TestDirectory, "Rabbit.yaml")));
+            using var model = Connection.Value.CreateModel();
+            model.ExchangeDeclare("TEST.ControlExchange", ExchangeType.Topic, durable: true);
         }
-
     }
 }
