@@ -23,7 +23,6 @@ namespace SmiServices.UnitTests.Common.Messages
         [OneTimeSetUp]
         public void OneTimeSetUp()
         {
-            TestLogger.Setup();
         }
 
         [OneTimeTearDown]
@@ -84,9 +83,9 @@ namespace SmiServices.UnitTests.Common.Messages
 
             try
             {
-                var exc = Assert.Throws<Exception>(() => new MessageHeader());
+                var exc = Assert.Throws<Exception>(static () => _ = new MessageHeader());
 
-                Assert.That(exc.Message, Is.EqualTo("Value must be set before use"));
+                Assert.That(exc?.Message, Is.EqualTo("Value must be set before use"));
             }
             finally
             {
